requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
ExifRead==3.5.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jpype1==1.6.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
konlpy==0.6.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.15.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
piexif==1.1.3
pillow==12.1.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.249
ultralytics-thop==2.0.18
urllib3==2.6.3
watchdog==6.0.0
